+++ src/Parser/Visitors/PauseAdderVisitor.php
